// src/utils/enums.ts
// Define a TypeScript enum for schema types
export enum SchemaTypes {
  STRING = 'string',
  NUMBER = 'number',
  BOOLEAN = 'boolean',
  ARRAY = 'array',
  OBJECT = 'object',
  INTEGER = 'integer',
}

// Define an enum for common formats
export enum SchemaFormats {
  FLOAT = 'float',
  DOUBLE = 'double',
  UUID = 'uuid',
  DATE_TIME = 'date-time',
}